+++ frontend/components/halamanutamacs/halamanutama.js
@@ -1,22 +1,13 @@
 import style from "./halamanutama.module.scss";
-import AdminCard from "./component/admincard/admincard";
 import Image from "next/image";
-import foto1 from "./component/assets/info1.jpg";
-import foto2 from "./component/assets/info2.jpg";
+import foto1 from "../../public/assets/info1.jpg";
+import foto2 from "../../public/assets/info2.jpg";
 
 const HalamanUtama = () => {
   return (
     <div className={style.utama}>
       <h1 className={style.title}>Selamat Datang Mr. Amron</h1>
       <hr style={{ width: "1200px" }} />
-      <br />
-      <div className={style.admincards}>
-        <h3 style={{ fontSize: "24px", fontWeight: "450" }}>
-          List Customer Service
-        </h3>
-        <AdminCard />
-      </div>
-      <br />
       <br />
       <div className={style.informasi}>
         <h3 style={{ fontSize: "24px", fontWeight: "450" }}>Informasi</h3>
