+++ frontend/components/sidebaradmin/sidebarn.js
@@ -1,6 +1,6 @@
 import style from "./sidebarn.module.scss";
 import Image from "next/image";
-import Logo from "../../../../../public/logo.png";
+import Logo from "../../public/logo.png";
 import HomeIcon from "@mui/icons-material/Home";
 import AccountBoxIcon from "@mui/icons-material/AccountBox";
 import InfoIcon from "@mui/icons-material/Info";
