--- frontend/pages/project/customerservice/component/halamanutama/component/admincard/admincard.module.scss
@@ -1,28 +1,0 @@
-.admincard {
-  margin-right: 10px;
-  margin-top: 10px;
-  margin-left: 10px;
-  display: grid;
-  grid-template-rows: 250px 250px;
-  grid-template-columns: minmax(0, 205px) minmax(0, 205px) minmax(0, 205px);
-  grid-gap: 30px;
-  min-width: 0;
-  min-height: 0;
-
-  .item {
-    padding: 5px;
-    overflow: hidden;
-    box-shadow: rgba(0, 0, 0, 0.24) 0px 3px 8px;
-    border-radius: 5px;
-    .foto {
-      margin-top: 5px;
-      max-height: 200px;
-    }
-    p {
-      font-size: 16px;
-      max-width: 100%;
-      max-height: 100%;
-      font-weight: 400;
-    }
-  }
-}
